+++ src/main/java/org/sciborgs1155/robot/Robot.java
@@ -15,17 +15,17 @@ import org.sciborgs1155.lib.failure.Fallible;
 import org.sciborgs1155.lib.failure.FaultBuilder;
 import org.sciborgs1155.lib.failure.HardwareFault;
 import org.sciborgs1155.robot.Ports.OI;
-import org.sciborgs1155.robot.commands.Autos;
+import org.sciborgs1155.robot.arm.Arm;
+import org.sciborgs1155.robot.arm.ArmConstants.Elbow;
+import org.sciborgs1155.robot.arm.ArmConstants.Elevator;
+import org.sciborgs1155.robot.arm.ArmConstants.Wrist;
+import org.sciborgs1155.robot.arm.ArmState;
+import org.sciborgs1155.robot.arm.ArmState.GamePiece;
+import org.sciborgs1155.robot.arm.ArmState.Goal;
+import org.sciborgs1155.robot.auto.Autos;
 import org.sciborgs1155.robot.drive.Drive;
-import org.sciborgs1155.robot.subsystems.Intake;
-import org.sciborgs1155.robot.subsystems.LED;
-import org.sciborgs1155.robot.subsystems.arm.Arm;
-import org.sciborgs1155.robot.subsystems.arm.ArmConstants.Elbow;
-import org.sciborgs1155.robot.subsystems.arm.ArmConstants.Elevator;
-import org.sciborgs1155.robot.subsystems.arm.ArmConstants.Wrist;
-import org.sciborgs1155.robot.subsystems.arm.ArmState;
-import org.sciborgs1155.robot.subsystems.arm.ArmState.GamePiece;
-import org.sciborgs1155.robot.subsystems.arm.ArmState.Goal;
+import org.sciborgs1155.robot.intake.Intake;
+import org.sciborgs1155.robot.intake.LED;
 import org.sciborgs1155.robot.vision.NoVision;
 import org.sciborgs1155.robot.vision.VisionIO;
