+++ src/main/java/org/sciborgs1155/robot/arm/NoElevator.java
@@ -1,18 +1,17 @@
-package org.sciborgs1155.robot.subsystems.arm;
+package org.sciborgs1155.robot.arm;
 
-import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.math.trajectory.TrapezoidProfile.State;
 import java.util.List;
 import org.sciborgs1155.lib.failure.HardwareFault;
 
-public class NoJoint implements JointIO {
+public class NoElevator implements ElevatorIO {
 
   @Override
   public void close() throws Exception {}
 
   @Override
-  public Rotation2d getRelativeAngle() {
-    return new Rotation2d();
+  public double getHeight() {
+    return 0;
   }
 
   @Override
@@ -32,20 +31,12 @@ public class NoJoint implements JointIO {
   public void stopMoving() {}
 
   @Override
-  public void setBaseAngle(Rotation2d baseAngle) {}
-
-  @Override
-  public Rotation2d getBaseAngle() {
-    return new Rotation2d();
+  public double getVoltage() {
+    return 0;
   }
 
   @Override
   public List<HardwareFault> getFaults() {
     return List.of();
-  }
-
-  @Override
-  public double getVoltage() {
-    return 0;
   }
 }
