+++ src/main/java/org/sciborgs1155/robot/arm/NoJoint.java
@@ -1,17 +1,18 @@
-package org.sciborgs1155.robot.subsystems.arm;
+package org.sciborgs1155.robot.arm;
 
+import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.math.trajectory.TrapezoidProfile.State;
 import java.util.List;
 import org.sciborgs1155.lib.failure.HardwareFault;
 
-public class NoElevator implements ElevatorIO {
+public class NoJoint implements JointIO {
 
   @Override
   public void close() throws Exception {}
 
   @Override
-  public double getHeight() {
-    return 0;
+  public Rotation2d getRelativeAngle() {
+    return new Rotation2d();
   }
 
   @Override
@@ -31,12 +32,20 @@ public class NoElevator implements ElevatorIO {
   public void stopMoving() {}
 
   @Override
-  public double getVoltage() {
-    return 0;
+  public void setBaseAngle(Rotation2d baseAngle) {}
+
+  @Override
+  public Rotation2d getBaseAngle() {
+    return new Rotation2d();
   }
 
   @Override
   public List<HardwareFault> getFaults() {
     return List.of();
+  }
+
+  @Override
+  public double getVoltage() {
+    return 0;
   }
 }
