+++ src/main/java/org/sciborgs1155/robot/arm/RealElevator.java
@@ -1,4 +1,4 @@
-package org.sciborgs1155.robot.subsystems.arm;
+package org.sciborgs1155.robot.arm;
 
 import static org.sciborgs1155.robot.Ports.Elevator.*;
 
@@ -15,7 +15,7 @@ import org.sciborgs1155.lib.constants.MotorConfig;
 import org.sciborgs1155.lib.failure.FaultBuilder;
 import org.sciborgs1155.lib.failure.HardwareFault;
 import org.sciborgs1155.robot.Constants;
-import org.sciborgs1155.robot.subsystems.arm.ArmConstants.Elevator;
+import org.sciborgs1155.robot.arm.ArmConstants.Elevator;
 
 public class RealElevator implements ElevatorIO {
