+++ src/main/java/org/sciborgs1155/robot/arm/RealWrist.java
@@ -1,4 +1,4 @@
-package org.sciborgs1155.robot.subsystems.arm;
+package org.sciborgs1155.robot.arm;
 
 import static org.sciborgs1155.robot.Ports.Wrist.*;
 
@@ -17,7 +17,7 @@ import org.sciborgs1155.lib.constants.MotorConfig;
 import org.sciborgs1155.lib.failure.FaultBuilder;
 import org.sciborgs1155.lib.failure.HardwareFault;
 import org.sciborgs1155.robot.Constants;
-import org.sciborgs1155.robot.subsystems.arm.ArmConstants.Wrist;
+import org.sciborgs1155.robot.arm.ArmConstants.Wrist;
 
 public class RealWrist implements JointIO {
