+++ src/main/java/org/sciborgs1155/robot/arm/Visualizer.java
@@ -1,4 +1,4 @@
-package org.sciborgs1155.robot.subsystems.arm;
+package org.sciborgs1155.robot.arm;
 
 import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.util.sendable.Sendable;
@@ -9,7 +9,7 @@ import edu.wpi.first.wpilibj.smartdashboard.MechanismRoot2d;
 import edu.wpi.first.wpilibj.util.Color;
 import edu.wpi.first.wpilibj.util.Color8Bit;
 import org.sciborgs1155.robot.Constants.Dimensions;
-import org.sciborgs1155.robot.subsystems.arm.ArmConstants.Elevator;
+import org.sciborgs1155.robot.arm.ArmConstants.Elevator;
 
 /** Visualization class specific for our charged up bot */
 public class Visualizer implements Sendable {
