+++ src/main/java/org/sciborgs1155/robot/vision/RealVision.java
@@ -1,6 +1,6 @@
 package org.sciborgs1155.robot.vision;
 
-import static org.sciborgs1155.robot.Constants.Vision.*;
+import static org.sciborgs1155.robot.vision.VisionConstants.*;
 
 import edu.wpi.first.apriltag.AprilTag;
 import edu.wpi.first.apriltag.AprilTagFieldLayout;
